test/integration-test/python-sdk-test/requirements.txt
nb-log==4.5
PyYAML==5.3.1
pytest==5.3.5
allure-pytest==2.8.9
